Extract the HandlerEnum into a dispatch module
