+++ src/handler/mod.rs
@@ -1,9 +1,9 @@
+pub mod dispatch;
 pub mod live_effect;
 pub mod live_looper;
 pub mod setup_looper;
 
 use crate::hmi::inputs::{Edge::Activate, InputEvent};
-use core::fmt;
 use defmt::*;
 use heapless::Vec;
 use midi_types::{MidiMessage, Note};
@@ -21,9 +21,7 @@ pub struct MidiMessages(MidiMessageVec);
 
 impl MidiMessages {
     pub fn push(&mut self, a: MidiMessage) {
-        if self.0.push(a).is_err() {
-            error!("failed pushing midi message")
-        };
+        self.0.push(a).unwrap();
     }
 
     pub fn clear(&mut self) {
@@ -60,39 +58,6 @@ pub trait Handler {
     fn leds(&mut self) -> &mut Leds;
 }
 
-pub enum HandlerEnum {
-    LiveEffect(self::live_effect::LiveEffect),
-    LiveLooper(self::live_looper::LiveLooper),
-    SetupLooper(self::setup_looper::SetupLooper),
-}
-
-impl Handler for HandlerEnum {
-    fn handle_human_input(&mut self, e: InputEvent) -> Actions {
-        match self {
-            HandlerEnum::LiveEffect(h) => h.handle_human_input(e),
-            HandlerEnum::LiveLooper(h) => h.handle_human_input(e),
-            HandlerEnum::SetupLooper(h) => h.handle_human_input(e),
-        }
-    }
-    fn leds(&mut self) -> &mut Leds {
-        match self {
-            HandlerEnum::LiveEffect(h) => h.leds(),
-            HandlerEnum::LiveLooper(h) => h.leds(),
-            HandlerEnum::SetupLooper(h) => h.leds(),
-        }
-    }
-}
-
-impl fmt::Debug for HandlerEnum {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            HandlerEnum::LiveEffect(_) => f.debug_tuple("LiveEffect").finish(),
-            HandlerEnum::LiveLooper(_) => f.debug_tuple("LiveLooper").finish(),
-            HandlerEnum::SetupLooper(_) => f.debug_tuple("SetupLooper").finish(),
-        }
-    }
-}
-
 const MAX_HANDLERS: usize = 8;
 pub type HandlerVec<H> = Vec<H, MAX_HANDLERS>;
 
@@ -101,6 +66,7 @@ pub struct Handlers<H: Handler> {
     current: usize,
 }
 
+/// The router (dispatcher) for human input and midi input
 impl<H> Handlers<H>
 where
     H: Handler,
@@ -158,6 +124,7 @@ where
     }
 }
 
+/// Construct empty Handlers
 impl<H> Default for Handlers<H>
 where
     H: Handler,
